Upper the warning level, made warnings fatal, isolated a bunch of warning disables to just LLVM, and fixed the resulting warnings.

+++ Source/AST/AST.h
@@ -102,7 +102,7 @@ namespace AST
 	{
 		const char* name;
 		std::vector<Variable> locals;
-		std::vector<uint32_t> parameterLocalIndices;
+		std::vector<uintptr_t> parameterLocalIndices;
 		FunctionType type;
 		UntypedExpression* expression;
